Created a MenuMeshLoader to load multiple collision meshes in-menu
